requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
requests
google-auth
google-auth-oauthlib
python-dotenv 